moved Noop back into main OpenTracing assembly
