+++ OpenTracing.sln
@@ -11,10 +11,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenTracing", "src\OpenTrac
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenTracing.Tests", "test\OpenTracing.Tests\OpenTracing.Tests.csproj", "{F82B4191-F9D1-4656-8585-840E6BA045DD}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenTracing.Noop", "src\OpenTracing.Noop\OpenTracing.Noop.csproj", "{53E3B733-D814-43B7-96B9-CD31500C309A}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenTracing.Noop.Tests", "test\OpenTracing.Noop.Tests\OpenTracing.Noop.Tests.csproj", "{366D9B7D-247C-4902-93D1-37E5CB49A6BC}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -29,14 +25,6 @@ Global
 		{F82B4191-F9D1-4656-8585-840E6BA045DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F82B4191-F9D1-4656-8585-840E6BA045DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F82B4191-F9D1-4656-8585-840E6BA045DD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{53E3B733-D814-43B7-96B9-CD31500C309A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{53E3B733-D814-43B7-96B9-CD31500C309A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{53E3B733-D814-43B7-96B9-CD31500C309A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{53E3B733-D814-43B7-96B9-CD31500C309A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{366D9B7D-247C-4902-93D1-37E5CB49A6BC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{366D9B7D-247C-4902-93D1-37E5CB49A6BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{366D9B7D-247C-4902-93D1-37E5CB49A6BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{366D9B7D-247C-4902-93D1-37E5CB49A6BC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -44,8 +32,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{60B9BDE0-8203-46C9-9373-E06FC82CECB3} = {CA88CFB0-F5DF-411B-B083-D54819E9414F}
 		{F82B4191-F9D1-4656-8585-840E6BA045DD} = {FC8B6803-6F66-4F1E-9034-5564B2D4FF9C}
-		{53E3B733-D814-43B7-96B9-CD31500C309A} = {CA88CFB0-F5DF-411B-B083-D54819E9414F}
-		{366D9B7D-247C-4902-93D1-37E5CB49A6BC} = {FC8B6803-6F66-4F1E-9034-5564B2D4FF9C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7DE89E0C-BFD0-4E2E-8093-56BE7E4F3820}
