--- src/OpenTracing.Noop/OpenTracing.Noop.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard1.0</TargetFrameworks>
-    <Description>OpenTracing NoOp implementation</Description>
-    <PackageTags>opentracing;distributed-tracing;tracing;logging</PackageTags>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(OS)'=='Windows_NT'">
-    <TargetFrameworks>$(TargetFrameworks);net45</TargetFrameworks>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\OpenTracing\OpenTracing.csproj" />
-  </ItemGroup>
-
-</Project>
