+++ src/OpenTracing/Noop/NoopScopeManager.cs
