+++ src/OpenTracing/Noop/NoopSpan.cs
