+++ src/OpenTracing/Noop/NoopTracer.cs
